profile , device 초안 주석 commit #수정1
departmentService 주석 처리 수정
